requirements.txt
south==1.0
selenium==2.44
psycopg2
fabric
django-geojson
django-colorful
Pillow

django-suit
sorl-thumbnail
